refactor data_sources file names
