+++ app/app.py
@@ -1,6 +1,6 @@
 from flask import Flask, jsonify, request
 
-from data_sources.datastore import get_call_history_records_by_interviewer_and_date_range, \
+from data_sources.datastore_data import get_call_history_records_by_interviewer_and_date_range, \
     get_call_history_report_status
 from functions.date_functions import validate_date
 from interviewer_call_pattern_report.report import get_call_pattern_records_by_interviewer_and_date_range
