+++ functions/call_history_functions.py
@@ -1,10 +1,10 @@
-from data_sources.blaise_api import get_list_of_installed_questionnaires, get_questionnaire_data
-from data_sources.cati_database import get_cati_call_history_from_database
-from data_sources.datastore import (
+from data_sources.cati_data import get_cati_call_history_from_database
+from data_sources.datastore_data import (
     get_call_history_keys,
     bulk_upload_call_history,
     update_call_history_report_status,
 )
+from data_sources.questionnaire_data import get_list_of_installed_questionnaires, get_questionnaire_data
 from models.call_history import CallHistory
